pyproject.toml
[tool.poetry]
name = "greenatom"
version = "0.1.0"
description = "chat API"
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
sqlalchemy = "^2.0.36"
aio-pika = "^9.4.3"
asyncpg = "^0.30.0"
phonenumbers = "^8.13.50"
sqlmodel = "^0.0.22"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
sqlalchemy-utils = "^0.41.2"
httpx = "^0.27.2"
backoff = "^2.2.1"
orjson = "^3.10.11"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
pyjwt = "^2.10.0"
websockets = "^14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
